// src/utils/enums.ts
export enum Roles {
  Leads = 'LEADS',
  Customer = 'CUSTOMER',
  FieldAgent = 'FIELDAGENT',
  UnderWriter = 'UNDERWRITER',
  Manager = 'MANAGER',
  Admin = 'ADMIN',
  FinanceManager = 'FINANCEMANAGER'
}